refactor(plugin): extract runCurationOp into a pure memory/curation-op module (#573 groundwork)
Recovered from a cc-glm worker whose worktree I deregistered before checking
for uncommitted work — the files survived on disk. Its approach is better than
the one I specced, so it is kept.

WHY: the #573 agent tool needs the same pin/unpin/retype/set_scope dispatch the
CLI has. Importing runCurationOp from cli/tr-cli.ts would drag the entire CLI
bin graph (pairing, onboarding, subgraph-store) into the in-process agent-tool
path. Extracting it to memory/curation-op.ts — a pure module whose only imports
are the execute* operations and their types — gives both call sites the same
normalizer with no transitive weight, and keeps the file scanner-clean by
construction (no env reads, no network).

tr-cli.ts now imports from there (1012 -> 915 lines); the two curation test
files were still importing runCurationOp from the old location and are
repointed (parseCurationCliArgs stays in tr-cli, runCurationOp + types come
from curation-op).

This is groundwork only — the agent tool itself is NOT yet wired. Remaining:
add curate to TrNativeMemoryDeps, register the tool, build PinOpDeps in
index.ts buildRecallDeps, and declare the tool name in openclaw.plugin.json
contracts.tools (or the loader silently drops it).

Verified: 98/98 test files pass; scanner 0 flags (177 files).
